requirements.txt
Flask==3.0.3
openai-whisper==20231117
transformers==4.43.3
torch==2.3.1
sentencepiece==0.2.0
